+++ engine/launcher/args/args.go
@@ -1,27 +1,29 @@
 package args
 
 import (
-	"github.com/kurtosis-tech/stacktrace"
 	"reflect"
 	"strings"
+
+	"github.com/kurtosis-tech/stacktrace"
 )
 
 const (
-	jsonFieldTag          = "json"
+	jsonFieldTag = "json"
 )
 
 // Fields are public for JSON de/serialization
 type EngineServerArgs struct {
-	ListenPortNum      uint16 `json:"listenPortNum"`
+	GrpcListenPortNum      uint16 `json:"grpcListenPortNum"`
+	GrpcProxyListenPortNum uint16 `json:"grpcProxyListenPortNum"`
 
-	LogLevelStr string	`json:"logLevelStr"`
+	LogLevelStr string `json:"logLevelStr"`
 
 	// So that the engine server knows its own version
 	ImageVersionTag string `json:"imageVersionTag"`
 
 	// The engine needs to know about this so it knows what filepath on the host machine to use when bind-mounting
 	//  enclave data directories to the API container & services that the APIC starts
-	EngineDataDirpathOnHostMachine string	`json:"engineDataDirpathOnHostMachine"`
+	EngineDataDirpathOnHostMachine string `json:"engineDataDirpathOnHostMachine"`
 
 	//The anonymized user ID for metrics analytics purpose `json:"metricsUserId"`
 	MetricsUserID string `json:"metricsUserId"`
@@ -30,11 +32,11 @@ type EngineServerArgs struct {
 	DidUserAcceptSendingMetrics bool `json:"didUserAcceptSendingMetrics"`
 }
 
-
 // Even though the fields are public due to JSON de/serialization requirements, we still have this constructor so that
 //  we get compile errors if there are missing fields
 func NewEngineServerArgs(
-	listenPortNum uint16,
+	grpcListenPortNum uint16,
+	grpcProxyListenPortNum uint16,
 	logLevelStr string,
 	imageVersionTag string,
 	engineDataDirpathOnHostMachine string,
@@ -42,7 +44,8 @@ func NewEngineServerArgs(
 	didUserAcceptSendingMetrics bool,
 ) (*EngineServerArgs, error) {
 	result := &EngineServerArgs{
-		ListenPortNum:                  listenPortNum,
+		GrpcListenPortNum:              grpcListenPortNum,
+		GrpcProxyListenPortNum:         grpcProxyListenPortNum,
 		LogLevelStr:                    logLevelStr,
 		ImageVersionTag:                imageVersionTag,
 		EngineDataDirpathOnHostMachine: engineDataDirpathOnHostMachine,
@@ -61,7 +64,7 @@ func (args EngineServerArgs) validate() error {
 	reflectVal := reflect.ValueOf(args)
 	reflectValType := reflectVal.Type()
 	for i := 0; i < reflectValType.NumField(); i++ {
-		field := reflectValType.Field(i);
+		field := reflectValType.Field(i)
 		jsonFieldName := field.Tag.Get(jsonFieldTag)
 
 		// Ensure no empty strings