+++ engine/docs/changelog.md
@@ -1,5 +1,9 @@
 # TBD
 
+### Features
+
+* Added GenericEngineClient interface to abstract grpc-web vs grpc-js Typescript backends
+
 # 1.11.1
 ### Changes
 * Kurtosis enclaves created via `KurtosisContext.CreateEnclave` default to the debug loglevel